气阀板更新
1.使用can,气阀板控制夹弹旋转电机,由gas_task里面控制
2.完成维特智能激光程序
3.调试了气阀板与主板通信
4.气阀控制变量名根据实际重命名

+++ rm_relay/ApplicationBSP/bsp_TOF.h
@@ -4,7 +4,7 @@
 #include "stdint.h"
 
 
-extern uint16_t TOF[2];
+extern uint8_t TOF[2];
 
 uint16_t TOF_GetData(uint8_t *receive_buff);
